iac/aws/cdk/cdk-project/requirements.txt
-e .
aws_cdk.aws_ec2